+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Move `process-event` from the Relay CLI into a standalone tool. ([#740](//github.com/getsentry/relay/pull/740))
 - Add the client SDK to session kafka payloads. ([#751](https://github.com/getsentry/relay/pull/751))
 - Add a standalone tool to document metrics in JSON or YAML. ([#752](https://github.com/getsentry/relay/pull/752))
+- Emit `processing.event.produced` for user report and session Kafka messages. ([#757](https://github.com/getsentry/relay/pull/757))
 
 ## 20.8.0